Rename reader to rateReader and add the reader and readerAt structures
