+++ internal/rate/rate.go
@@ -9,7 +9,7 @@ import (
 	"golang.org/x/time/rate"
 )
 
-type reader struct {
+type rateReader struct {
 	r              io.Reader
 	ra             io.ReaderAt
 	l              *rate.Limiter
@@ -18,26 +18,8 @@ type reader struct {
 	once           sync.Once
 }
 
-// NewReader create a limit io.Reader that wrap the real io.Reader.
-// The bytesPerSecond must be greater than defaultBufSize of io.Reader.
-func NewReader(r io.Reader, bytesPerSecond int64) io.Reader {
-	if bytesPerSecond <= 0 {
-		return r
-	}
-	return newReader(r, nil, bytesPerSecond)
-}
-
-// NewReaderAt create a limit io.ReaderAt that wrap the real io.ReaderAt.
-// The bytesPerSecond must be greater than defaultBufSize of io.ReaderAt.
-func NewReaderAt(ra io.ReaderAt, bytesPerSecond int64) io.ReaderAt {
-	if bytesPerSecond <= 0 {
-		return ra
-	}
-	return newReader(nil, ra, bytesPerSecond)
-}
-
-func newReader(r io.Reader, ra io.ReaderAt, bytesPerSecond int64) *reader {
-	return &reader{
+func newRateReader(r io.Reader, ra io.ReaderAt, bytesPerSecond int64) *rateReader {
+	return &rateReader{
 		r:              r,
 		ra:             ra,
 		l:              rate.NewLimiter(1, 1),
@@ -45,19 +27,19 @@ func newReader(r io.Reader, ra io.ReaderAt, bytesPerSecond int64) *reader {
 	}
 }
 
-func (r *reader) Read(p []byte) (n int, err error) {
+func (r *rateReader) Read(p []byte) (n int, err error) {
 	return r.call(func() (n int, err error) {
 		return r.r.Read(p)
 	})
 }
 
-func (r *reader) ReadAt(p []byte, off int64) (n int, err error) {
+func (r *rateReader) ReadAt(p []byte, off int64) (n int, err error) {
 	return r.call(func() (n int, err error) {
 		return r.ra.ReadAt(p, off)
 	})
 }
 
-func (r *reader) call(f func() (n int, err error)) (n int, err error) {
+func (r *rateReader) call(f func() (n int, err error)) (n int, err error) {
 	r.once.Do(func() {
 		r.l.Reserve()
 	})
